improved naming based on review suggestions
